Basic proxybot message routing working w/ xmlrpc
Since mod_admin_extra lets ejabberdctl add and remove roster items, it seemed worth trying to switch to use ejabberd_xmlrpc for registering/unregistering users and adding/deleting roster items, rather than relying on in-band registration and hacky modifications to ejabberd itself. This seems to be working pretty well, and hopefully ejabberdctl can now serve all of our needs and we can limit the scope of changes we need to make to ejabberd.

The most basic proxybot functionality is now working - the bot gets created, logs in, and adds itself to each user's roster and pretends to be the other user. The bots then route messages between the users, so they can have the beginnings of a normal conversation.