Add proxybot->hostbot adhoc communication channel
The proxybot clients need to be able to tell the hostbot component when a new user joins or leaves a conversation or when the proxybot moves to an active or retired state, so that the hostbot can keep track of the current state in a database. This commit adds the communication channel necessary for this using XMPP's ad-hoc command (XEP-0050), which are support by both SleekXMPP and ejabberd.

Notes:
	- I had some issues with threading and the proxybot disconnecting after finishing the retire command, so I needed to fire a custom event with a threaded handler so that the proxybot wouldn't disconnect until after the ad-hoc communication was done.
	- I'm not thrilled with how long/unclear all of the new function names are, but I'm not sure what's better at the moment.
	- I'm also not thrilled with how similar the functions are for the different events, but I don't yet see a way to abstract them that is actually any clearer.
	- The hostbot isn't making the actual modifications to the database yet, but that should be a pretty easy next step.