Add observers to proxybots on create_friendship
When a friendship is created, either user might be in a conversation that the other should be aware of and have in his/her conversations list. So, on create_friednship, look for the active proxybots for which one user is a participant but not the other, and add the latter user as an observer to the former's participant object.

There are also a few other small fixes here. Next, I need to make it so that conversations are appropriately removed from the conversations list on delete_friendship.