Fix observer bug after /new_ or /del_friendship
So, this one is particiularly insidious, and is maybe only easy to spot with a new user that I am simultaneously testing and setting up with friendships for the first time.

Observers are only changed when the social graph changes. Since this is relatively uncommon, before we were not querying the database to fetch a *new* observer list when a proxybot was activated, but only when a friendship was created or deleted. This is fine, but then when we added or removed observers when the social graph was changed, we were only sending that notification to the *active* proxybots, and not the *idle* ones. But then when those idle conversations were becoming active, they did not have the correct observer lists, and someone either would have been able to see a conversation they shouldn't have seen, or didn't see a conversation they should have seen.