Fix that observer bug, for real this time
See 749f58a7a7ae2fa2633a50cb40b54a05c0a3c13b for a description of this bug, but my solution was wrong. When I created a new friendship it was filling
 each other's buddy list with proxybots from the other user's buddy list, and gave them all the other buddy's  name, which is totally broken.

This should fix it. Rather than try to update the observer lists for idle proxybots, just don't fetch the observer lists in the first place until the proxybot is activated.