Add timestamps to conversation topics
This was much more annoying than expected. It seems like there was a bug in the Bot class's __setattr__ method in which self.__dict__.has_key('_attrExample__initialised') was always False, and thus none of the special __setattr__ stuff was ever getting called. It now doesn't rely on that check to see if the function is initialized, and only handles the special cases with a modified function.

Some day I might try to figure out why it wasn't working as expected, but this is good for now. http://code.activestate.com/recipes/389916-example-setattr-getattr-overloading/ is the example I used that didn't do what I expected.