Add partial support for message & /command logging
So, before I wasn't really keeping track of what /commands were being used, but was only storing the output messages. This isn't really good enough if I want to weight things such as /invite differently in the vine-graph edge calculations, or if I want to figure out what commands users are using most often, or etc. I really needed to be logging almost everything.

To solve this, it made sense to begin to split apart the message and /command logging. Under this scheme, logged_messages will have the messages sent to users, log_recipients will keep track of who received what messages, and logged_commands will keeep track of the received (and correctly formatted) comamnds.

It seemed important, here, to be able to associate the output messages with the command that triggered them in the database. To facilate this the messages need the id of the logged_command, which made the SlashCommand stuff *much* more complicated. It seemed simplest to insert the logged_command when the args were transformed successfully, and then return the id from there both to handle_msg, which may need to reply, as well as the action method for that command, which may send messages of its own (since both types of messages need to be associated with the logged_command that prompted them). I'm not really happy with how this came out, since it feels like spaghetti, but I don't know how else to structure things while still enabling /command code-re-use.

Next, I'm going to fix/add the logging of normal messages. The tables should be able to account for all of the cases I have in a reasonable and structured way.