Add logging of messages (paired with /commands)
Now all sent messages are properly logged by the leaf, as well as received /commands.

For valid /commands, messages that are sent as a result of the command are stored with the corresponding logged_command_id, so that they can be linked together later.

Invalid /commands are still stored, along with the error message that is sent back.

For messages not sent to a vinebot, or messages sent to an invalid vinebot, the recevied message won't always be logged, and the response won't be linked. I'm going to add that next.

Note that there's a lot of passing arrond of the logged_command_id between the different functions. I don't really love this, but anything that's broadcasting a message needs it, so I don't know how else to do it.