Re-organzie methods and files for new leaf design
The new version of the leaves will support having multiple leaves connected to ejabberd at the same subdomain. If different leaves nearly-simultaneously receive messages for the same vinebot that require database writes, then one has to wait until the other is finished. To accomplish this I'll use MySQL's application locks and database transactions, and it makes sense to group more of this functionality together in the Bot class (which I'll rename to Vinebot soon).

Each leaf, however, is still a single-threaded python process, so it's fine if each only has one database connection and one XMLRPC user that it uses for whichever vinebot is associated with the current request/message. Rather than have the leaf pass itself to the vinebot, it's cleaner to only pass the database and XMLRPC objects, and to encapsulate the setup and basic commands for those things in separate classes.