Add check on startup to see if other leaves online
Use MySQL's application locks to register, in as light a way as possible, that there's a leaf connected. Then, if there are no other leaves, we know we have to send the initial presence stanzas for all of the vinebots. If there are other leaves online, then we don't ned to do that.