WIP commit, some progress on create_ & delete_edge
Everything is a mess right now, so VERY work in progress, but I think the basics of the create_edge and delete_edge functions are there.

In addition to the Vinebot class, there is now an Edge class and a User class, to ecapsulate data and database queries for different types of objects. They're essentially custom data models, without the possibly-inefficient and opaque SQL.

I ended up with a decent amount of symmetry between the two functions, which makes me happy, but there's lots of broken/extraneous stuff in the new classes. It's going to be a while before this runs :-/