Add/fix a ton of stuff for asymmetric refactor
Please please forgive me for a commit this big. I started the refactor, and didn't see a good stopping point. And this is not a good stopping point, but it's better than when everything is working.

The handle_presence_*, handle_msg, /del_edge, /new_edge, /del_user, and /new_user are mostly working, as well as the startup/shutdown behavior. There are bugs all over the place, but add_participant and remove_participant are mostly working, with rosters and presences also mostly working.