Fix a nasty mutability bug w/ immutable frozensets
I was using the set.pop() method to get the remaining element from a vinebot's list of participants so that I would have the name of the last user that I wanted to remove from that vinebot when de-activating it. This was buggy, since pop() modifies the set, so then the vinebot had nothing to remove. I'm now using immutable frozensets, and this shouldn't be an issue any more.