Fix previously-unnoticed bug in broadcast_message
Suddenly, broadcast_message seemed to be sending messages back to the sender. This was because I had failed to implement __ne__ when I implemented __eq__.

Also, clean up the edge stuff when activating vinebots in handle_msg.

And pass the full user objects as recipients to log_message, since it's cleaner.