Add the /topic command for vinebots
Note that /topic now explicitly supports cases when the conversation is not already active: if the other user is online, it makes the vinebot active, and if the other user is offline, it just clears the topic. To do this I refactored out the vinebot activation roster/presence stuff into a leaf method.