Finish /del_user admin cmd for edges & vinebots
Lots of stuff here, but mostly, when a user is getting deleted, you need to remove it from active conversations, and delete both incoming and outgoing edges. To do this I needed to add a couple more database query functions, and generalize some of the edge deletion code that's also used by /del_edge. I also went through and made FetchedEdge calls more explicit about which of the optional parameters they were using, since I had a nasty bug where a from_user and to_user were switched in a SQL query and I wanted to be sure to be explicit.