Use diff mysql connections for each vinebot lock
This fixes the issue mentioned in 237b0f78514bda1550ea719fc70589c5ba3badae, in which MySQL's application locks (using get_lock) are not released if a second lock is acquired by the same connection. Sometimes the leaf needs multiple vinebot locks (and always needs its leaf lock), so, for lack of a more elegant solution, this now uses a separate MySQL connection for each vinebot lock. It tries to keep track of and re-use these connections in a lightweight pool, and hopefully it never needs too many at a time (i.e. more than a few).

I feel like the new database classes are really messy, and that this is probably not the most elegant way to manage connections and query commands. This does seem like the least code, however, and it does seem like it requires the least changes to everything else, so it'll do for now.

Some links about this shortcoming of MySQL:
http://bugs.mysql.com/bug.php?id=1118
http://dev.mysql.com/worklog/task/?id=1159
http://board.issociate.de/thread/99039/GETLOCKstrtimeout-behaviour.html