Don't delete vinebots from db b/c of foreign keys
Both the commands and messages table have foreign keys to vinebots, so we can't delete them, and we should keep them around anyway so that we can group conversation histories properly. It might be worth knowing the jiduser for the log view, perhaps for synchronization with other logs from something like Adium? hmmm.