Fix presence handling when vinebots are activated
Before, if someone was away, they would come back briefly is a user sent them a message, and then later go away again, since the person wasn't actually there. This was annoying, and it is now smarter about the presence stanzas that it sends on behalf of an active vinebot.

Also, commit changes in the shared submodule.