Specify python eggs directory (see note, ARGH)
I was getting an error from gevent's monkey patching about the Python egg cache being set to /root/.python-eggs, which supervisord didn't have permission to write to. I've never heard of a Python egg cache beore, and was a bit surprised that gevent was trying to use a system cache instead of a virtualenv-specific one, but it seemed easiest to just take advantage of the supervisord configuration setting and set the PYTHON_EGG_CACHE environment variable.

Except, ARGH, this didn't work. The file path must be in quotes, since otherwise for some reason it's just interpreted as "/", sigh.