Fix XMLRPC calls w/ custom Transport, scrap gevent
This was a nightmare. Wow. A gist with my test script and more info/error messages: https://gist.github.com/4471424

So, the new, Chef-managed Vagrant VM is running Ubuntu 12.04, which comes with Python 2.7.3. This broke the add_rosteritem and delete_rosteritem calls I was making in ejabberdctl.py, and I was getting CannotSendRequest and ResponseNotReady errors from httplib. It still worked in Python 2.6.5, but I wasn't able to install 2.6 alongside 2.7 because it was missing libraries like libssl and zlib, and I wasn't able to figure out why.

I tried lots of different ways of spawning asynchronous requests, including eventlet and geventhttpclient, but couldn't get around the error. At first I was only getting the errors when the whole thing was within a SleekXMPP component that I told to .connect(), but later I figured out that that was only because that was delaying the exiting of my test script, and a time.sleep(2) accomplished the same thing.

I finally started to dig into xmlrpclib.py, and noticed that they let you define a custom Transport. I started to write one, and after more digging into how httplib.py handles connections and responses, I ended up with a Transport that closes the connection instead of waiting for a response. I think the connection is being closed anyway by the standard Transport class, so it shouldn't be any less efficient. This is a nice solution, since there's no threading overhead, and I don't have to monkey.patch my entire XMPP component, which I was kinda worried was introducing race conditions with the database state. All of the gevent-related code has been removed, and it's no longer installed by the cookbook.

I did get a good response in the gevent group. Shaun Lindsay (from Meebo! now at Google) suggested that the problem was because I was re-using the same ServerProxy object for multiple concurrent requests. This makes sense, and I should have been able to figure this out. The thread in the gevent group has more good details: https://groups.google.com/forum/?fromgroups=#!topic/gevent/sNS8QiEd3RQ

Lesson learned: If a library isn't doing what you want it to do, read the source.