Refactor data bag attributes for vine-chef updates
Now, everything is stored directly in the node.run_state hash, so that all the different recipies have access to it.

This also uses the updated deploy key data bag for fetching from the private git repo.

Finally, since we got rid of the node attributes, we have to add the history items here rather than in vine-chef. They probably belong here. but really, we need a resource to do this for us so that code isn't copy-pasted.