Use gevent.spawn for add/delete_rosteritem again
It seems like most of our XMLRPC requests using the custom fire-and-forget transport were snappy, but if we had to make many of them (30+ for my user account), some of them were slow, and as a result the entire modification of a conversation's participants was taking 2 seconds or longer, which is much too long.

So, now we're using a hybrid of what we added in de7b613c2a889b82bf0a2ae8245331f13a7e55f8 and what we were doing before that. Hopefully this is fast enough, and that the monkey patching behaves!