Add leaf /commands for managing account invites
They are:
	/hide_invite <code>
	/show_invite <code>
	/new_invite <sender>
	/del_invite <code>
	/list_invites <sender>

Nothing too interesting going on here. Added a new set of Invite classes that mirrored the others to handle the database queries, and the commands themselves are pretty self-explanatory.

Would love to have all of the commands abstracted somehow into another service, since I feel like they're starting to clutter the actual functioning of the LeafComponent... but that is for another day.