Add random sync_roster in handle_presence_available
So that if there are errors in a user's roster, we have a probability of fixing them every time the user logs in. Note that this probability is higher for users with lots of vinebots in their roster, since the leaves receive more presence_availables, but these users are also more likely to have roster errors, so I guess that's ok.