Allow mixed case /command args, make lower() later
If the user uses a mixed-case token arg (i.e. for a username) then it won't properly split up the later arguments in the transform_args function, since it forces the tokens to lower and then can't find the new lowercase arg_token with the partition in the arg_string. I'm not entirely sure why it was like this in the first place, but this should be better.

Also, now, the User class constructors are smart about making sure the usernames they accept are made lowercase.