Don't always register InsertedUsers
We need this for the /tweet_invite command, since the invites need a valid recipient, but that user doesn't have a password yet and shouldn't be able to sign in.

Note that unregistered users should still be set to active. Otherwise, the won't get returned for the admin's /edges and /purge_user commands, and they'll throw exceptions. There are probably other errors too.