Let leaf_component use celery tasks from web repo
It would be helpful if the leaves can use score_edges in particular to update a user's edges periodically, either on sign-in or a /block or /unblock command or etc. To make that work, I'm using the vine-web repo as a module within the vine-xmpp repo (although not as a git-submodule? which might be helpful, but oh well) that Chef sets up for me. I also need to add a bunch of stuff to the constants template that was only used by the web repo before, since that same template now gets rendered for both repos.