Make sure to clear cached user presence in Adium
Ugh this was hard to figure out.

So, I want to put the active conversations in a separate buddy list group, so that Adium users can create notifications for when users in that group sign on, so that they can know when there are conversations happening without having their buddy list open.

After struggling with the "contact signs on" event in Adium not working, and digging through the Adium source and asking questions in #adium-devl, and being told I should be using the "contact is seen" event instead, I went back to Sleek and played around there.

I eventually discovered that removing the account from Adium and then re-adding it made it work the first time, but not latter times, and then I realized that Adium was perhaps caching the isOnline status for users not on the buddy list! Explicitly sending a presence unavailable before removing the roster item fixes it :)

(This should have been happening before, maybe? But whatever.)