Divide roster into Conversations+Contacts groups
Now there are two Vine-managed roster groups: Vine Conversations for all the active conversations that you're a part of and that you can see, and Vine Contacts for all the online contacts. This was pretty easy to do in the ejabberdctl commands, but was a bit tricky to get to work with /sync_roster, and there might still be some bugs there... shipping it anyway :)

The main reason for doing this is so that Adium users can set group-specific event alerts. If they listen for the "Contact signs on" event, then they can get notified whenever a new conversation starts that they can see. This was inspired by a conversation with Christina, who doesn't look at her buddy list a lot, but still wants ambient awarenss of what's going on around her. There's some detail in caf9752f1e5d9e1161a488933700bc4cf6226e8c about what was needed to make this work. Fortunately, because your contacts are already online and in the "seen" state when you send them a message to start a conversation, you won't get notified for conversations you start.

I also made it so that active conversations have the nickname "you & jabberwocky" instead of just "jabberwocky", so that people have visibility into what other conversations are still active that other people can see and join. Perhaps this will encourage people to leave conversations, but it also might encourage people to continue active conversations, or remind people of messages they never responded to. And either way, if there are fewer idle conversations, then there won't be as much of the negative UX when someone joins an empty or half-empty room.

Note that this is how the groups were managed in a very early version (maybe back in the days of proxybot python processes, lol). One big difference is that if you're talking to someone, they don't have a separate inactive contact, so you can't inadvertently start multiple conversations with the same person. I could re-implement this, but I'm not sure it's better, and it requires touching more of the code.