When roster changes, send available w/ unavailable
ZOMG THIS WAS A NIGHTMARE. For some reason, Adium seems to improperly handle presences with type='unavailable' immediately preceding (and I think following) roster pushes. So, for example, this XML:

	<presence xmlns='jabber:client' from='vinebot_uqx3y8ocvbwv268fy9adff1z4@leaves.dev.vine.im' to='jabberwocky@dev.vine.im/Stevens-MacBook-Air' type='unavailable'/>

	<iq xmlns='jabber:client' from='jabberwocky@dev.vine.im/Stevens-MacBook-Air' to='jabberwocky@dev.vine.im/Stevens-MacBook-Air' type='set' id='push'>
		<query xmlns='jabber:iq:roster'>
			<item jid='vinebot_uqx3y8ocvbwv268fy9adff1z4@leaves.dev.vine.im' name='alice' subscription='both'>
				<group>Vine Contacts (Dev)</group>
			</item>
		</query>
	</iq>

	<iq type='result' id='push'/>

Resulted in a weird state for the corresponding vinebot where the status read "Offline" but the availability was still set to green, so it still appeared as online in the roster, which is no good. UGH. I tried a ton of stuff, including removing and then re-adding the roster item, but couldn't make it work for hours.

I'm guesssing this is an actual Adium issue, based on http://www.cocoaforge.com/viewtopic.php?p=117300 and http://forums.cocoaforge.com/viewtopic.php?t=24520 and talking with kbotc in #adium-devel, but I really don't want to be debugging Adium, especially since a fix won't go out till the next release and won't fix this bug for me/my users now.

The workaround is to send a presence available immediately before the presence unavailable in cases where there is a roster change. I'm guessing this re-associates the JID with the new item in Adium's presence cache, or something weird like that, but it should be harmless otherwise, and isn't much extra traffic.