Fix phantom roster item bug for inactive vinebots
Jorge accidentally added a contact to a roster of a different Adium account, so still saw it as online. When he tried to send a message to that contact to test it, I think Adium defaulted to sending it from his Vine account, and I wasn't checking to make sure that the message the activated a vinebot was coming from an edge_user, so Aaron and I received the phantom ping from Jorge even though he wasn't in the conversation, and we couldn't respond to him.