Update admin /commands to allow multi-use invites
There are corresponding changes for this in vine-web and vine-chef.

Basically, I need to be able to use invites more than once, so I can send a single email to an email list with a single invite link that everyone can use, but I also need a way to restrict how many times that invite does get used in case it gets shared too publicly. This requires changes to the datase scheme and the Invite classes, and a couple places in the leaf component, but nothing too fancy or complicated.