Only activate vinebots if both users say something
When a user sends a message to a vinebot, rather than immediately add both edge_users as participants, instead only add them (and thus "activate" the vinebot) if the other user has sent a message to that vinebot in the last 10 minutes. This way conversations that aren't necessarily active won't appear in the Vine Conversations group for either those users or any of the observers, so observers can be more sure that they're joining an active conversation rather than intruding on a conversation that hasn't even started.