Add helpbot.py and onboarding flow to leaves
This corresponds to changes in vine-chef to implement a super-basic onboarding flow for new users. It keeps track in the database of how far through the flow the user has progressed, and when the leaf receives a presence available from a user, it checks to see if the user has been onboarded and pings the helpbot, which then responds accordingly. I'll add a lot more complexity to this later, but want to do something simple first.