Fix major bugs in and clean up onboarding bot
Ugh, I deployed broken code and accidentally sent messages to Jorge and I and maaaaybe everyone else, but I hope they didn't have edges to @vineim and the messages didn't go through.

Anyway, I cleaned up the code for this a bit. The biggest change is that it now sends the continue onboarding command from the vinebot corresponding to the edge between the users. The bug before was that the command was sent from whatever vinebot sent the presences, which is no good. Since I was having to fetch that vinebot anyway, it was easier to just do once in the leaf, and then update the helpbot to expect messages to come from the vinebot to which they correspond.