Remove chatstates from observers
I think it already sends chatstates to all observers, even those that are offline, which is fine, but then if they're offline they'll send back errors and I was then trying to broadcast those errors, which was causing the leaf to do way too much work. At least, that's what it seemed like, I'm not totally sure, and those chatstates aren't important enough to be worth investigating.