Fix MAJOR helpbot bugs that caused infinite loop
When I made some /commands start with /** when they were sent privately as responses to the sender, and the broadcast ones start with ***, I didn't update the helpbot, so as a result the helpbot tried to respond to messages it shouldn't have.

There was also a NameError with how the vinebot lock was getting released, since the Vinebot wasn't defined if the helpbot erroneously sent the message directly to the leaf, so that was throwing an error. I might have this mistake elsewhere vinebot locks are getting released, but I'll worry about it in the leaf later I guess.

I think these two bugs caused an infinite loop in prod for like 16 hours. I'm going to have to delete a bunch of rows from the database to clean it up, but I'll take a snapshot of the current state.