Update presences sent when participant is removed
I'm trying to speed up the /leave and /kick commands, which are taking way too long way too much of the time, and are often causing other attempts to lock that vinebot to time out, which just slows down everthing even more.

Maybe this is happening because these commands sent out a flurry of presences, and the user clients would respond, and then the leaves would get stuck handling them? And I don't think I need to be sending those presences in many cases, since the status of the vinebot will not actually have changed.

The one presence we might need is the probes sent when a conversation ends, and the vinebot again has to adopt the presences of the two edge users, but we should move that into remove_participant from user_left, because that only needs to happen if there are edge users in the first place who have an inactive vinebot that is getting replaced.

Hopefully nothing breaks :-/ I want my own XMPP server sighhh.