Fix chatstate bug with offline users
It was re-broadcasting chatstates with type="error", which was causing offline users to send the composing states back.

This may or may not fix the bug where vinebots get stuck in a typing state. If not, I might add the following to send_idle_presences, along with some other stuff:
	msg = self.Message()
	msg['typ`e'] = 'chat'
	msg['chat_state']
	self.broadcast_message(active_vinebot.everyone)