Fix deploy_wrapper version issue again (see 153463738f5e07516e4631def1fc25b0ff7400e8)
I don't really understand why this is happening, but the error message looks like this:

	2013-05-20T07:05:20+00:00] INFO: HTTP Request Returned 412 Precondition Failed: {"unsatisfiable_run_list_item":"(vine_xmpp >= 0.0.0)","non_existent_cookbooks":[],"most_constrained_cookbooks":["Package deploy_wrapper"],"message":"Unable to satisfy constraints on cookbook deploy_wrapper, which does not exist, due to run list item (vine_xmpp >= 0.0.0). Run list items that may result in a constraint on deploy_wrapper: [(vine_xmpp = 1.0.5) -> (deploy_wrapper = 0.0.4)]"}

I tried to fix it by specifying versin >= 0.0.2, but that didn't work. Then I went to the communtiy site and downloaded 0.0.2, rather than 0.0.4 (not sure why this happened in the first place), and put that on my Hosted Chef, and it seems to have worked. Weird, whatever.