Put topics in vinebot nicknames, else participants
Re-factor Vinebot statuses and nicknames to allow the topic to be in the nickname and the participant list to be in the status, and vice-versa (as things were before). This required a bunch of changes to various aspects of the Vinebot, including how topics and last activity were stored, in order to handle the variety of ways in which the information can be presented in the roster with minimal code duplication. Some specifics:

	- Build status strings in vinebot, not leaf.

	- Send presences whenever the participant list changes, since the status might change then too.

	- Be more explicit about checking if the `topic is not None`.

	- Use "me" as the pronoun for referring to the user instead of "you", and put this in a constant so it's easy to change later.

	- Move marshal_nick and unmarshal_nick from the sync_roster method into the AbstractVinebot class as static methods, and spell them correctly.

	- Fix bugs in sync_roster, including one where, if a rosteritem had the wrong nick, it would be deleted and then re-added due to the way the set differences work, but then sometimes the delete would execute after the add, which is no good and leaves the roster in an incorrect state, and is inefficient since the add would have changed the nick just fine on its own.

	- Other sync_roster cleanup.

	- Clean up lots of whitespace in vinebot.py