Suspend notifications when users disconnect, /join
This is based on feedback from tdjoe and jonshea at Foursquare, who were sad when they came back to their computer and saw unread messages, only to discover that they were automated notifications saying that someone had left the conversation.

/join notifications sent to rooms with more than three participants and all disconnect notifications are now "suspended" automatically. This means they don't get delivered to the room until some other message is also getting sent.

I'm flagging suspended messages in MySQL by setting the sent_on column to '0000-00-00 00:00:00' (note that I initially wanted to use NULL, but my timestamp column automatically set it to the current time; I'm now taking advantage of that feature to consolidate other aspects of this commit). Now, whenever a message is going to be broadcast to a conversation, it checks to see if there are any suspended messages for that vinebot with the special zero timestamp. If there are, it sends them to the participants who would have received both messages – that way, if someone is /invited, they won't see what happened before they arrived, even if the message was suspended until the /invite notification was broadcast. After broadcasting the suspended messages to the conversation, it updates the recipients and the sent_on timestamp so that the message doesn't need to get unsuspended again.

Note the actual time that the events occurred is still getting logged by the leaf with "[join]" and "[offline]", and those don't contain information about the user or vinebot. But the disconnects are logged in ejabberd, and the /join command is still logged as a /command, so I don't think I'm actually losing any data.

Also, note that I needed to update the MySQL user privileges in the vine_shared cookbook in the vine-chef repository, so that the leaf can update the messages table and delete from the recipients table – see https://github.com/lehrblogger/vine-chef/commit/1a89453907d66d052dc36455444cb53602b502e5.