Add "/party user_1,user_2 /topic a topic" command
It invites the listed users to a conversation with the specified topic, but they only receive future messages and are considered participants if they reply and join the conversation. Intended to be useful if you want to talk about something but don't want to interrupt anyone specifically.

There's a tiny bit of hacky stuff in letting vinebots have only one user (the person sending the message), and in how users who received the invite message are considered observers of the vinebot, but all things considered it was a pretty clean change :)