Add script to subscribe users to MailChimp list
This is currently a one-off script to move over all of my existing users, so that I can easily send an email to all of the ones that didn't sign in. That said, I think I'm going to need more sophisticated email management tools in the future, so I think it'll be good to have as a starting point.

Notes:
	I added a new ejabberdctl command to mod_admin_extra itself, since I couldn't find an API for the last activity info that I saw in the web dashboard: https://github.com/lehrblogger/ejabberd-modules/commit/7021082784ae22e378b7d74b4d243573b7ed1560
	I added the email field to the user object, hopefully this doesn't break anything.
	I needed to make minor changes to this cookbook and to the ejabberd cookbook.