Fix python-eggs directory permissions warning
The error was:
	UserWarning: /home/vagrant/.python-eggs is writable by group/others and vulnerable to attack when used with get_resource_filename. Consider a more secure location (set with .set_extraction_path or the PYTHON_EGG_CACHE environment variable).
I'm not sure why it started happening, but this fixes it. I tried to figure out if it was something with how gevent was getting installed and did LOTS of trial and error (note that I can probably use pypi now, but it's not worth changing), but then I discovered the gevent file in the ~/.python-eggs directory wasn't even getting put there at all until python ran for the first time, even if in a virtual env. I guess gevent needs a global copy or something? Weird. Anyway, specifying the permissions on that directory seems to fix it, sigh.